Iimplementation of mathematics controller
