+++ api/models/joke.py
