+++ api/requirements.txt
@@ -1,4 +1,4 @@
 fastapi==0.85.0
 uvicorn==0.18.3
-pytest==7.1.3
+pytest==7.1.2
 pymongo==4.2.0
