+++ api/routes/joke.py
