+++ api/schemas/joke.py
